Refactored Utils Module for V1 and V2
